pass script to bash in case /bin/bash is executable, even when started with sh
We want the script to work in environments w/o bash (in particular: OpenWrt busybox ash)
so the interpreter line requests /bin/sh. But we DO want to run it under bash on systems
that do have bash (because then, sh mode is likely not to be compatible with what we need
which is tuned to work for openwrt busybox ash).